pyproject.toml
[tool.poetry]
name = "timeline"
version = "0.1.0"
description = ""
authors = ["JD Bothma <jbothma@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
slack-bolt = "^1.13.0"
requests = "^2.27.1"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
timeline = "timeline.command_line:main"